pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polybet"
version = "0.1.0"
description = "Polymarket Sports Auto-Analyst"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.7",
  "typer>=0.12",
  "rich>=13.7",
  "tenacity>=8.2",
  "python-dotenv>=1.0",
  "diskcache>=5.6"
]

[project.optional-dependencies]
dev = ["pytest>=8.2"]

[project.scripts]
polybet = "polybet.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
